[PAGOPA-1143] chore: refactored package for ECS log classes
